+++ README.md
@@ -1,2 +1,5 @@
 # B31210_MSCS11_OOP
 CSC8101 Advanced Object Oriented Programming Exam
+
+#Refugee Food Bank Management System 
+It is a platform designed to streamline food bank operations for refugees and displaced families. The system ensures fair and timely access to essential food supplies, manages food inventory, coordinates food distribution, and tracks beneficiaries and donors. The system also allows for the transparent management of donations and distributions.
